Squash merge branch 'wip/serde' into 'feat/serde'
