+++ engine/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "libfortivo"
-version = "1.0.0"
+version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-serde = { version = "1.0.219", features = ["derive"] }
+serde = "1.0.219"
 thiserror = "2.0.12"
