+++ engine/src/error.rs
@@ -1,26 +1,29 @@
-use thiserror::Error;
-use serde::ser;
 use std::fmt::Display;
+use serde::{de, ser};
+use thiserror::Error;
 
 /// Type alias for errors in this crate
-pub type FortivoResult<T> = Result<T, FortivoError>;
+pub(crate) type FortivoResult<T> = Result<T, FortivoError>;
 
-/// The main error type used by this crate, it implements the [`std::error::Error`] trait
+/// Main error type used by libfortivo
 #[derive(Error, Debug)]
 pub enum FortivoError {
-	#[error("Encountered an IO error")]
-	IO(#[from] std::io::Error),
-	#[error("Is clock correctly set?")]
-	Time(#[from] std::time::SystemTimeError),
-	#[error("Encountered an error trying to serialize")]
-	Serialize(String),
-	#[error("Arca name is too long, maximum allowed is 512 bytes")]
-	ArcaNameTooLong
+    #[error("Encountered an IO error")]
+    IO(#[from] std::io::Error),
+    #[error("Encountered an error trying to serialize")]
+    Serialize(String),
+    #[error("Encountered an error trying to deserialize")]
+    Deserialize(String),
 }
 
-// Implement serde's Error trait so this error type can be used with it
 impl ser::Error for FortivoError {
-	fn custom<T: Display>(msg: T) -> Self {
-		Self::Serialize(msg.to_string())
-	}
+    fn custom<T: Display>(msg: T) -> Self {
+        Self::Serialize(msg.to_string())
+    }
+}
+
+impl de::Error for FortivoError {
+    fn custom<T: Display>(msg: T) -> Self {
+        Self::Deserialize(msg.to_string())
+    }
 }
